refactor: move schemas to own file
